feat: use latest-weekly per default
The correlation calibration is only done weekly. Therefore, the Pong
demo loads the latest-weekly calibration. In the sable/experimentally
deployed notebooks this can lead to an mismatch between the calibration
and the software state since not specially tagged calibrations are used.
By using the weekly calibration per default (locally) and making sure to
tag/symlink weekly-calibrations for ebrains-stable and
ebrains-experimental we can make sure that calibration and code stay in
agreement.

Change-Id: I8f28d51d72d85db17530cba7b2d7c49d427b7872